Moved all homne page components to home folder
