+++ app/home/components/footer.tsx
