+++ app/home/components/hero.tsx
