+++ app/layout.tsx
@@ -2,10 +2,10 @@ import './global.css';
 import type { Metadata } from 'next';
 import { GeistSans } from 'geist/font/sans';
 import { GeistMono } from 'geist/font/mono';
-import { Navbar } from './components/nav';
+import { Navbar } from './home/components/nav';
 import { Analytics } from '@vercel/analytics/react';
 import { SpeedInsights } from '@vercel/speed-insights/next';
-import Footer from './components/footer';
+import Footer from './home/components/footer';
 import { ThemeProvider } from './components/theme-switch';
 import { metaData } from './config';
 import { Space_Grotesk, Roboto_Flex } from 'next/font/google';
